requirements.txt
pip==22.1.2
setuptools==63.1.0
wheel==0.37.1
streamlit
python-dotenv
wheel
aniso8601==8.0.0
appdirs==1.4.4
argopt==0.6.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
coloredlogs==14.0
gunicorn==20.0.4
humanfriendly==8.2
img2pdf==0.3.6
importlib-metadata==1.7.0
itsdangerous==1.1.0
ocrmypdf==10.2.0
pdfminer.six==20200517
pikepdf==1.17.0
Pillow==7.2.0
pluggy==0.13.1
pycparser==2.20
pycryptodome==3.9.8
pyocr==0.7.2
python-pdfbox==0.1.8
pytz==2020.1
reportlab==3.5.44
requests==2.24.0
six==1.15.0
sortedcontainers==2.2.2
sqlitedict==1.6.0
tqdm==4.47.0
urllib3==1.25.9
Werkzeug==2.2.2
joblib==1.1.0
opencv-python
matplotlib

# Requirements for the training pipeline
pytorch_lightning
transformers
albumentations
mlflow
pytesseract
